feat: move css to composables. Added option to manually choose css modules.
